requirements.txt
tweepy
googletrans
uwsgi
gunicorn
python-dotenv